requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.1
